Only update blocked user meta if user isn't already blocked
A previous commit used the return of `bp_update_user_meta()`, updating the blocked user meta key, to determine if the expiration time should be updated and the user's sessions should be destroyed. This was to prevent unnecessary attempts to write to the database. However, `bp_update_user_meta()` would returns `false` if the meta value doesn't change. This prevented updates to user expiration times, and caused certain admin screens to display an when a user was already blocked.