Remove caching on `tba_bp_get_blocked_user_ids()`
This function is only called once per page load, and only while loading the blocked users list table. It's also not a heavy database call. Rather than add the complication of caching (and cache invalidation) for very little gain, it's easier to remove the code. As needs change, we may need to add on caching or a transient, but for now, we're okay.