Better user meta handling in `BPBU_User::block` an `BPBU_User::unblock`
Both the `update_metadata()` and `delete_metadata()` functions used by their BP user meta wrappers have checks to prevent unnecessary writes to the database. We should rely on those.

On top of that, `update_metadata()` will return `false` when the meta value isn't changing. This was causing issues in the yet-to-be-published test suite.