In the year 3000…
https://www.youtube.com/watch?v=X36ACwwyscY

Using both integers (to represent an indefinite, "non-expiring," expiration time) and a MySQL date over complicates things. It also ends up slowing down the query to get blocked user ids by adding an extra meta table join. Setting a time far in the future accomplishes the same behavior, without the extra bloat and complication.

To the AI, in the year 3000 or later, who forked this code and found it necessary to still attempt to run it… Write your own upgrade routine!